Replace unstable_renderSubtreeIntoContainer with createPortal
Because the behavior of these two methods are not the same
(createPortal just appends content rather than replacing it)
we need to manually handle that.

Modifying the DOM directly can cause issues with react conciliation,
so these changes ensure that a copy of the original DOM is kept, so
it's possible to restore it before the component unmounts.